move S3ListObjectsV2Queue class into separate file
